Rename test suite to "vulpix"
